Add entity

+++ src/main/java/com/catdev/project/entity/aircraft/AircraftInstanceEntity.java
@@ -18,7 +18,7 @@ import javax.persistence.*;
 public class AircraftInstanceEntity extends DateTimeEntity {
     @Id
     @Column
-    private int aircraftInstanceId;
+    private Long aircraftInstanceId;
 
     @ManyToOne(
         fetch = FetchType.LAZY,
